+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Verb Changelog
 ## master
 - Renamed `verb-mode-prefix-map` to `verb-command-map` (existing `init.el` configs must be updated).
+- Signal an error when using `verb-set-var` if no variables have been previously set with `verb-var`.
 
 ## **2.1.0** - 2020-01-24
 - Verb can now be used from Babel source blocks.